[WIP][capture] #34: split directions for libpcap
First steps towards using the standard libpcap library. At the moment,
it relies on BPF filters "inbound" and "outbound".

The current code is _not_ stable, since it fails when running on MacOS,
where apparently, the BPF filters aren't supported.
